Moved mapping helpers to outside module
Addresses https://github.com/tschaub/mock-fs/pull/304#discussion_r461959380
